+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 - sound: Replace pyaudio with miniaudio
 - lib.hardware.tispeech: Avoid using all 0s and 1s for nearest energy lookups
 - utils.find_iter: Fix binary searching
+- lib.platforms.director.unlock_dir_file: Improve robustness, remove dependency on parser
 
 0.9.0 - 2021-01-14
 ==================